Fix code styles and javadoc
