+++ data/daniel.txt
@@ -1,3 +1,3 @@
-D | 0 | return book | 2/12/2019 1800
-E | 0 | meeting | 2/12/2019 2200 | 2/12/2019 2300
+E | 1 | meeting | 2/12/2019 2200 | 2/12/2019 2300
+D | 0 | hi | 3/12/2028 1111
 T | 0 | read book
