+++ src/main/java/exception/MissingArg.java
@@ -1,6 +1,6 @@
-package Exception;
+package exception;
 
-public class MissingArg extends Exception{
+public class MissingArg extends Exception {
     public MissingArg(String message) {
         super(message);
     }
