+++ src/main/java/exception/WrongIndex.java
@@ -1,6 +1,6 @@
-package Exception;
+package exception;
 
-public class WrongIndex extends Exception{
+public class WrongIndex extends Exception {
     public WrongIndex(String message) {
         super(message);
     }
