+++ src/main/java/exception/WrongKeyword.java
@@ -1,6 +1,6 @@
-package Exception;
+package exception;
 
-public class WrongKeyword extends Exception{
+public class WrongKeyword extends Exception {
     public WrongKeyword(String message) {
         super(message);
     }
