+++ src/test/java/task/TodoTest.java
@@ -1,4 +1,4 @@
-package Task;
+package task;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
